pyproject.toml
[tool.poetry]
name = "moodle-scraper"
version = "0.1.0"
description = ""
authors = ["ia rigby <marvinzem@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.18.1"
python-dotenv = "^1.0.1"
gcsa = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
